pyproject.toml
[project]
name = "console_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.120.3",
    "greenlet>=3.2.4",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "bcrypt==4.0.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "email-validator>=2.0.0",
    "colorama>=0.4.6",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
]
